Struct nom_grapheme_clusters::parse::Tag

pub struct Tag<'slice, 'seg>(pub &'slice [&'seg str]);

A type usable as tag for a parser without having to create a proper Span.

Tuple Fields
0: &'slice [&'seg str]

A sequence of segment contents.

Implementations

impl<'slice, 'seg> Tag<'slice, 'seg>

pub fn len(self) -> usize

Returns the length of the tag in segments/grapheme clusters.

pub fn into_fn<T, E>(self) -> impl FnMut(T) -> IResult<T, T, E> + 'slice + 'seg where
'slice: 'seg,
T: InputTake + Compare<Self>,
E: ParseError<T>,

Converts this tag parsed into a function (also a parser).

pub fn segments(self) -> SegmentContents<'slice, 'seg>

Returns an iterator over the contents of segments of this tag.