Trait nom_grapheme_clusters::source::SourceIndex

pub trait SourceIndex: Debug {
    type Output: ?Sized;

    fn get<'src>(&self, src: &'src Source) -> Option<&'src Self::Output>;

    fn index<'src>(&self, src: &'src Source) -> &'src Self::Output { ... }
}

An index on a source code.

Required Associated Types

type Output: ?Sized

Output of the indexing operation.

Required Methods

fn get<'src>(&self, src: &'src Source) -> Option<&'src Self::Output>

Indexes the source code and returns None if out of bounds.

Provided Methods

fn index<'src>(&self, src: &'src Source) -> &'src Self::Output

Indexes the source code and panics if out of bounds.