Struct nom_grapheme_clusters::span::Span

pub struct Span { /* private fields */ }

A span (a range) in the source code.

It can be created with the constructor Span::new, or with Source::full_span.

Implementations

impl Span

pub fn new(start: Location, length: usize) -> Self

Creates a new span given the start location and length.

Panics

Panics if length is too big, i.e. start.position() + length > start.source().len().

pub fn from_range(start: Location, end: Location) -> Self

Creates a new span from a range of locations, where end is exlusive.

Panics

Panics if start and end point to different sources, or if start is past beyond end.

pub fn start(&self) -> Location

The start location of this span.

pub fn end(&self) -> Location

The end location of this span.

pub fn len(&self) -> usize

The length of this span in string segments.

pub fn source(&self) -> &Source

The source code object this span refers to.

pub fn as_str(&self) -> &str

Gets the string this span includes as a whole.

pub fn content(&self) -> SpanContent

Creates a type that, when displayed, shows the span contents, rather than location.

pub fn expand_lines(&self) -> Span

Expands this span in order to contain the whole lines the original span contains.

pub fn try_slice<R>(&self, range: R) -> Option<Self>where
R: RangeBounds<usize>,

Slices this span to the given range. Returns None if the range is invalid.

pub fn segments(&self) -> Segments

Creates an iterator over located grapheme cluster segments, namely LocatedSegments.

pub fn indexed_segments(&self) -> IndexedSegments

Creates an IndexedSegments iterator, which yields a tuple of position and a LocatedSegment in that position. Note that this is just for convenience with e.g. nom, LocatedSegment already contains its position, and so Segments can be used.