Module nom_grapheme_clusters::parse

Structs

Tag

A type usable as tag for a parser without having to create a proper Span.

Functions

alpha0

Recognizes zero or more UTF-8 alphabetic segments, possibly with diacritics.

alpha1

Recognizes one or more UTF-8 alphabetic segments, possibly with diacritics.

alphanumeric0

Recognizes zero or more UTF-8 alphanumeric segments, possibly with diacritics.

alphanumeric1

Recognizes one or more UTF-8 alphanumeric segments, possibly with diacritics.

any_segment

Recognizes any grapheme cluster/segment.

ascii_alpha0

Recognizes zero or more ASCII alphabetic segments without diacritics.

ascii_alpha1

Recognizes one or more ASCII alphabetic segments without diacritics.

ascii_alphanumeric0

Recognizes zero or more ASCII alphanumeric segments without diacritics.

ascii_alphanumeric1

Recognizes one or more ASCII alphanumeric segments without diacritics.

ascii_numeric0

Recognizes zero or more ASCII numeric segments without diacritics.

ascii_numeric1

Recognizes one or more ASCII numeric segments without diacritics.

char_alpha0

Recognizes zero or more UTF-8 alphabetic segments without diacritics.

char_alpha1

Recognizes one or more UTF-8 alphabetic segments without diacritics.

char_alphanumeric0

Recognizes zero or more UTF-8 alphanumeric segments without diacritics.

char_alphanumeric1

Recognizes one or more UTF-8 alphanumeric segments without diacritics.

char_numeric0

Recognizes zero or more UTF-8 numeric segments without diacritics.

char_numeric1

Recognizes one or more UTF-8 numeric segments without diacritics.

crlf

Recognizes the sequence '\r\n'.

digit0

Recognizes zero or more digits in the given base. ASCII characters 0-9. a-z, A-Z are considered digits, depending on the base.

digit1

Recognizes one or more digits in the given base. ASCII characters 0-9. a-z, A-Z are considered digits, depending on the base.

newline0

Recognizes zero or more linefeed ('\n') ASCII characters.

newline1

Recognizes one or more linefeed ('\n') ASCII characters.

numeric0

Recognizes zero or more UTF-8 numeric segments, possibly with diacritics.

numeric1

Recognizes one or more UTF-8 numeric segments, possibly with diacritics.

space0

Recognizes zero or more ASCII spaces.

space1

Recognizes one or more ASCII spaces.

whitespace0

Recognizes zero or more unicode whitespace graphemes.

whitespace1

Recognizes one or more unicode whitespace graphemes.