Struct nom_grapheme_clusters::source::Source

pub struct Source { /* private fields */ }

A source code object, such as read from a file. Cloning this object results in simply incrementing a reference counter, thus sharing the source code object.

Implementations

impl Source

pub fn new<S0, S1>(name: S0, contents: S1) -> Selfwhere S0: Into<Box<str>>, S1: Into<Box<str>>,

Creates a new source code object given its name and its contents.

Contents are rearranged as grapheme clusters.

pub fn name(&self) -> &str

The (file) name of the source.

pub fn len(&self) -> usize

The length the source.

pub fn contents(&self) -> &str

The contentss of the source.

pub fn seg_byte_indices(&self) -> SegmentByteIndices<'_>

Iterator over the segment indices of the source, where indices are in terms of bytes.

pub fn newline_indices(&self) -> NewlineIndices<'_>

Iterator over the newline indices of the source, where indices are in terms of segments/grapheme clusters.

pub fn get<I>(&self, indexer: I) -> Option<&I::Output>where I: SourceIndex,

Indexes this source. It can be a single usize or a range of usize. Indices are given in terms of grapheme clusters/segments.

pub fn full_span(&self) -> Span

Returns a span covering the whole source code.