Module nom_grapheme_clusters::span

This module provides ways of tracking ranges (spans) in the source code.

Structs

IndexedSegments

Iterator over segments of a Span which also yield postion for convenience. Created by Segments::indexed, Span::indexed_segments or SpanContent::indexed_segments. Double-ended and sized.

Segments

Iterator over located segments of a Span. Created by Span::segments or SpanContent::segments, as well via IntoIterator trait. Double-ended and sized.

Span

A span (a range) in the source code.

SpanContent

A type that, when displayed, shows the span contents, rather than location.

Symbol

A type for metadata associated with a span (“spanned data”).

Traits

Spanned

Types that have a span associated.